Moving effects frame to separate window
