+++ gui/showwindow.cpp
@@ -7,7 +7,6 @@
 #include "chasewizard.h"
 #include "configurationwindow.h"
 #include "controlwindow.h"
-#include "effectsframe.h"
 #include "presetsframe.h"
 #include "sceneframe.h"
 #include "visualizationwindow.h"
@@ -58,11 +57,9 @@ ShowWindow::ShowWindow(std::unique_ptr<DmxDevice> device) :
 	createMenu();
 	
 	_presetsFrame.reset(new PresetsFrame(*_management, *this));
-	_effectsFrame.reset(new EffectsFrame(*_management, *this));
 	_sceneFrame.reset(new SceneFrame(*_management, *this));
 
 	_notebook.append_page(*_presetsFrame, "Presets");
-	_notebook.append_page(*_effectsFrame, "Effects");
 	_notebook.append_page(*_sceneFrame, "Timeline");
 
 	_box.pack_start(_notebook);
