+++ gui/windowlist.h
@@ -6,26 +6,42 @@
 #include <memory>
 #include <vector>
 
+template<typename T>
 class WindowList
 {
 public:
-	void Add(std::unique_ptr<Gtk::Window> window)
+	void Add(std::unique_ptr<T> window)
 	{
 		window->signal_hide().connect([&]() { onHideWindow(window.get()); });
 		_list.emplace_back(std::move(window));
 	}
 	
-	const std::vector<std::unique_ptr<Gtk::Window>>& List() const { return _list; }
+	const std::vector<std::unique_ptr<T>>& List() const { return _list; }
 	
-private:
-	void onHideWindow(Gtk::Window* window)
+	T* GetOpenWindow(class FolderObject& object) const
 	{
-		auto iter = std::find_if(_list.begin(), _list.end(),
-			[window](const std::unique_ptr<Gtk::Window>& elem) { return elem.get() == window; } );
-		_list.erase(iter);
+		for(auto& window : _list)
+		{
+			if(&window->GetObject() == &object)
+				return window.get();
+		}
+		return nullptr;
 	}
 	
-	std::vector<std::unique_ptr<Gtk::Window>> _list;
+private:
+	void onHideWindow(T* window)
+	{
+		for(auto iter=_list.begin(); iter!=_list.end(); ++iter)
+		{
+			if(iter->get() == window)
+			{
+				_list.erase(iter);
+				break;
+			}
+		}
+	}
+	
+	std::vector<std::unique_ptr<T>> _list;
 };
 
 #endif
